sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conduit-sdk"
version = "0.1.0"
description = "Python SDK for building Exchange-compatible applications with the Conduit operator"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Conduit Contributors"}
]
keywords = ["kubernetes", "operator", "nats", "jetstream", "conduit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "nats-py>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
kubernetes = [
    "kubernetes>=28.0.0",
]

[project.urls]
Homepage = "https://github.com/tonyd33/conduit"
Documentation = "https://github.com/tonyd33/conduit/blob/master/sdk/python/README.md"
Repository = "https://github.com/tonyd33/conduit"
Issues = "https://github.com/tonyd33/conduit/issues"

[tool.setuptools]
packages = ["conduit_sdk"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
